chatbot fix
